+++ crates/wdl-lint/src/rules/unused_doc_comments.rs
@@ -21,14 +21,26 @@ use crate::TagSet;
 const ID: &str = "UnusedDocComments";
 
 /// Creates a diagnostic for a misplaced doc comment.
-fn unused_doc_comment_diagnostic(comment_span: Span, target_span: Option<Span>) -> Diagnostic {
-    let diagnostic = Diagnostic::note("unused doc comment")
+fn unused_doc_comment_diagnostic(
+    comment_span: Span,
+    target_span: Option<Span>,
+    valid_target: bool,
+    floating: bool,
+) -> Diagnostic {
+    let mut diagnostic = Diagnostic::note("unused doc comment")
         .with_rule(ID)
-        .with_highlight(comment_span)
-        .with_fix(
+        .with_highlight(comment_span);
+
+    if valid_target {
+        assert!(floating);
+        diagnostic =
+            diagnostic.with_help("doc comments must be attached to the item to be recognized");
+    } else {
+        diagnostic = diagnostic.with_fix(
             "if this is a non-doc comment, replace the leading `##` with `#`; otherwise move this \
              comment so it documents the intended item",
         );
+    }
 
     if let Some(target_span) = target_span {
         diagnostic.with_label(
@@ -70,6 +82,9 @@ const VALID_SYNTAX_KINDS_FOR_DOC_COMMENTS: &[SyntaxKind] = &[
     SyntaxKind::BoundDeclNode,
 ];
 
+/// [`SyntaxKind`]s that are allowed to have floating doc comments.
+const VALID_SYNTAX_KINDS_FOR_FLOATING_COMMENTS: &[SyntaxKind] = &[SyntaxKind::VersionStatementNode];
+
 /// Determines whether the SyntaxNodeOrToken is a valid target for a doc
 /// comment.
 fn valid_target_for_doc_comment(doc_comment_target: &SyntaxElement) -> bool {
@@ -89,14 +104,43 @@ fn valid_target_for_doc_comment(doc_comment_target: &SyntaxElement) -> bool {
     VALID_SYNTAX_KINDS_FOR_DOC_COMMENTS.contains(&kind)
 }
 
+/// The element targeted by a doc comment.
+struct DocCommentTarget {
+    /// The target element.
+    element: SyntaxElement,
+    /// Whether the comment is detached from the target.
+    floating: bool,
+}
+
 /// Finds the first non-trivia [`SyntaxElement`] in the comment's siblings
 /// to determine what this doc comment is targeting.
-fn search_siblings_for_doc_comment_target(comment: &Comment) -> Option<SyntaxElement> {
+fn search_siblings_for_doc_comment_target(comment: &Comment) -> Option<DocCommentTarget> {
     let mut next = comment.inner().next_sibling_or_token();
+    let mut floating = false;
     while let Some(sibling) = next {
         next = sibling.next_sibling_or_token();
-        if !sibling.kind().is_trivia() {
-            return Some(sibling);
+        match &sibling {
+            SyntaxElement::Node(_) => {
+                return Some(DocCommentTarget {
+                    element: sibling,
+                    floating,
+                });
+            }
+            SyntaxElement::Token(t) => match t.kind() {
+                SyntaxKind::Whitespace => {
+                    let lines = t.text().chars().filter(|c| *c == '\n').count();
+                    if !floating {
+                        floating = lines > 1;
+                    }
+                }
+                SyntaxKind::Comment => {}
+                _ => {
+                    return Some(DocCommentTarget {
+                        element: sibling,
+                        floating,
+                    });
+                }
+            },
         }
     }
     None
@@ -132,44 +176,80 @@ fn get_span_of_first_token_for_syntax_element(element: &SyntaxElement) -> Span {
 }
 
 impl UnusedDocCommentsRule {
+    /// Collects all doc comments in the block containing `comment`.
+    ///
+    /// Returns the span of the entire comment block.
+    fn collect_consecutive_doc_comments(&mut self, comment: &Comment) -> Span {
+        let mut next = comment.inner().next_sibling_or_token();
+        let mut span_end = comment.span().end();
+        while let Some(sibling) = next {
+            next = sibling.next_sibling_or_token();
+            if sibling.kind() == SyntaxKind::Whitespace {
+                let lines = sibling
+                    .as_token()
+                    .unwrap()
+                    .text()
+                    .chars()
+                    .filter(|c| *c == '\n')
+                    .count();
+                if lines > 1 {
+                    break;
+                }
+                continue;
+            }
+            if let Some(continued_comment) =
+                sibling.as_token().and_then(|t| Comment::cast(t.clone()))
+            {
+                self.skip_count += 1;
+                span_end = continued_comment.span().end();
+                continue;
+            }
+            break;
+        }
+
+        Span::new(comment.span().start(), span_end - comment.span().start())
+    }
+
     /// Produces an unused doc comment diagnostic for the doc comment block
     /// starting at `comment`. Updates `skip_count` along the way.
     fn lint_next_doc_comment_block(
         &mut self,
         diagnostics: &mut Diagnostics,
         comment: &Comment,
-        target_span: Option<Span>,
+        target: Option<DocCommentTarget>,
     ) {
-        let mut next = comment.inner().next_sibling_or_token();
-        let mut span_end = comment.span().end();
-        while let Some(sibling) = next {
-            next = sibling.next_sibling_or_token();
+        let (mut target_span, floating) = target
+            .as_ref()
+            .map(|target| {
+                (
+                    Some(get_span_of_first_token_for_syntax_element(&target.element)),
+                    target.floating,
+                )
+            })
+            .unwrap_or((None, false));
+        let valid_target = target
+            .as_ref()
+            .is_some_and(|t| valid_target_for_doc_comment(&t.element));
+        let valid_floater = target
+            .as_ref()
+            .is_some_and(|t| VALID_SYNTAX_KINDS_FOR_FLOATING_COMMENTS.contains(&t.element.kind()));
 
-            if sibling.kind() == SyntaxKind::Whitespace {
-                continue;
-            }
-
-            if let Some(continued_comment) =
-                sibling.as_token().and_then(|t| Comment::cast(t.clone()))
-                && continued_comment.kind() == CommentKind::Documentation
-            {
-                self.skip_count += 1;
-                span_end = continued_comment.span().end();
-                continue;
-            } else {
-                diagnostics.add(unused_doc_comment_diagnostic(
-                    Span::new(comment.span().start(), span_end - comment.span().start()),
-                    target_span,
-                ));
-                return;
-            }
+        let block_span = self.collect_consecutive_doc_comments(comment);
+        if valid_target && (!floating || valid_floater) {
+            return; // Valid doc comment
         }
 
-        // If the doc comment block extends to the end of the token stream,
-        // we won't add a diagnostic above. Add one here.
+        // Floaters targeting nodes that don't allow floaters should be linted
+        // separately, without targeting the node below them.
+        if floating && !valid_floater {
+            target_span = None;
+        }
+
         diagnostics.add(unused_doc_comment_diagnostic(
-            Span::new(comment.span().start(), span_end - comment.span().start()),
+            block_span,
             target_span,
+            valid_target,
+            floating,
         ));
     }
 }
@@ -276,22 +356,13 @@ impl Visitor for UnusedDocCommentsRule {
             diagnostics.add(unused_doc_comment_diagnostic(
                 comment.span(),
                 Some(get_span_of_first_token_for_syntax_element(&target)),
+                false,
+                false,
             ));
             return;
         }
 
         let target = search_siblings_for_doc_comment_target(comment);
-        if target
-            .as_ref()
-            .is_none_or(|t| !valid_target_for_doc_comment(t))
-        {
-            self.lint_next_doc_comment_block(
-                diagnostics,
-                comment,
-                target
-                    .as_ref()
-                    .map(get_span_of_first_token_for_syntax_element),
-            );
-        }
+        self.lint_next_doc_comment_block(diagnostics, comment, target);
     }
 }